feat: re-organize sample implementation directories WPB-19280
Change the group in sample-java from 'org.example' to 'com.wire.integrations'
